file renames in lite.text_generation
